alx_travel_app/listings/requirements.txt
django>=3.0,<5.0
djangorestframework>=3.12,<4.0
django-cors-headers>=3.7,<4.0
drf-yasg>=1.20,<2.0
django-environ>=0.8,<1.0
mysqlclient>=2.0,<3.0
celery>=5.2,<6.0
rabbitmq-server>=3.8,<4.0

